Fixed the structure in the backend.
